+++ agency/experimental/ndarray/ndarray.hpp
@@ -1,12 +1,16 @@
 #pragma once
 
 #include <agency/detail/config.hpp>
+#include <agency/detail/requires.hpp>
+#include <agency/detail/default_shape.hpp>
 #include <agency/experimental/ndarray/ndarray_ref.hpp>
 #include <agency/detail/utility.hpp>
 #include <agency/memory/allocator/detail/allocator_traits.hpp>
 #include <agency/detail/iterator/constant_iterator.hpp>
 #include <utility>
 #include <memory>
+#include <iterator>
+
 
 namespace agency
 {
@@ -58,9 +62,10 @@ class basic_ndarray
 
     __agency_exec_check_disable__
     template<class Iterator,
-             class = typename std::enable_if<
-               !std::is_convertible<Iterator,shape_type>::value
-             >::type>
+             // XXX this requirement should really be something like is_input_iterator<InputIterator>
+             __AGENCY_REQUIRES(
+               std::is_convertible<typename std::iterator_traits<Iterator>::value_type, value_type>::value
+             )>
     __AGENCY_ANNOTATION
     basic_ndarray(Iterator first, Iterator last)
       : basic_ndarray(agency::detail::shape_cast<shape_type>(last - first))
@@ -276,7 +281,7 @@ class basic_ndarray
 
 
 template<class T, size_t rank, class Alloc = std::allocator<T>>
-using ndarray = basic_ndarray<T, point<std::size_t,rank>, Alloc>;
+using ndarray = basic_ndarray<T, agency::detail::default_shape_t<rank>, Alloc>;
 
 
 } // end experimental