+++ testing/bulk_then/multiple_results.cpp
@@ -15,7 +15,7 @@ void test()
     auto fut = agency::make_ready_future<int>(policy.executor(), 7);
 
     auto f = agency::bulk_then(policy(10),
-      [](typename execution_policy_type::execution_agent_type& self, int& return_me)
+      [](typename execution_policy_type::execution_agent_type&, int& return_me)
       {
         return return_me;
       },
@@ -37,7 +37,7 @@ void test()
     auto fut = agency::make_ready_future<void>(policy.executor());
 
     auto f = agency::bulk_then(policy(10),
-      [](typename execution_policy_type::execution_agent_type& self)
+      [](typename execution_policy_type::execution_agent_type&)
       {
         return 7;
       },
@@ -61,7 +61,7 @@ void test()
     int val = 13;
 
     auto f = agency::bulk_then(policy(10),
-      [](typename execution_policy_type::execution_agent_type& self, int& past_arg, int val)
+      [](typename execution_policy_type::execution_agent_type&, int& past_arg, int val)
     {
       return past_arg + val;
     },
@@ -85,7 +85,7 @@ void test()
     int val = 13;
 
     auto f = agency::bulk_then(policy(10),
-      [](typename execution_policy_type::execution_agent_type& self, int val)
+      [](typename execution_policy_type::execution_agent_type&, int val)
     {
       return val;
     },
@@ -109,7 +109,7 @@ void test()
     int val = 13;
 
     auto f = agency::bulk_then(policy(10),
-      [](typename execution_policy_type::execution_agent_type& self, int& past_arg, int& shared_arg)
+      [](typename execution_policy_type::execution_agent_type&, int& past_arg, int& shared_arg)
     {
       return past_arg + shared_arg;
     },
@@ -133,7 +133,7 @@ void test()
     int val = 13;
 
     auto f = agency::bulk_then(policy(10),
-      [](typename execution_policy_type::execution_agent_type& self, int& shared_arg)
+      [](typename execution_policy_type::execution_agent_type&, int& shared_arg)
     {
       return shared_arg;
     },